# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG pipeline API — document ingestion, vector search, and LLM-powered chat
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Document {
  id: number;
  filename: string;
  originalName: string;
  status: DocumentStatus;
  chunkCount: number;
  createdAt: string;
  /** @nullable */
  errorMessage: string | null;
}

export interface DocumentFileUpload {
  file: Blob;
}

export interface Conversation {
  id: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Source {
  documentId: number;
  documentName: string;
  chunkIndex: number;
  text: string;
  score: number;
}

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  sources: Source[];
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface Stats {
  documentCount: number;
  readyDocumentCount: number;
  totalChunks: number;
  conversationCount: number;
  messageCount: number;
}