# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG pipeline API — document ingestion, vector search, and LLM-powered chat
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: number;
  filename: string;
  originalName: string;
  status: DocumentStatus;
  chunkCount: number;
  createdAt: string;
  /** @nullable */
  errorMessage: string | null;
}