# lib/api-zod/src/generated/types/documentFileUpload.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG pipeline API — document ingestion, vector search, and LLM-powered chat
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentFileUpload {
  file: Blob;
}